requirements.txt
langchain
InstructorEmbedding
torch
tqdm
sentence_transformers
pypdf
chromadb
wheel
